// src/lib/types.ts
export interface SiteSettings {
  id: string;
  companyName: string;
  companyEnName: string;
  phone: string;
  phone2: string;
  whatsapp: string;
  email: string;
  address: string;
  workingHours: string;
  twitter: string;
  instagram: string;
  linkedin: string;
  youtube: string;
  snapchat: string;
  tiktok: string;
  facebook: string;
}